+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [maintenance] Update all dependencies
 - [maintenance] Prune unneeded dependencies
+- [new] Add `mode` option to Story parameters. Most useful for its `warn` option, which will print errors detected for a story but not fail the test suite.
 
 ## 6.2.0 (2022-09-29)